feature-timetable: timetable

+++ scripts/app/auth/auth.interceptor.js
@@ -1,10 +1,14 @@
 angular.module('betterTimetable')
-    .factory('authExpiredInterceptor', function ($rootScope, $q, $injector, localStorageService) {
+    .factory('authExpiredInterceptor', function ($rootScope, $q, $injector, localStorageService, $location) {
         return {
             responseError: function (response) {
                 // session has expired
                 if (response.status === 401 ||  response.status === 403 || response.status === 419) {
-                    localStorageService.remove('isAuthenticated');
+                    if(localStorageService.get('isAuthenticated')){
+                        localStorageService.remove('isAuthenticated');
+                        Materialize.toast('Sorry, your session timed out', 4000);
+                        $location.path("/login");
+                    }
                 }
                 return $q.reject(response);
             }
